fix: Login state
