+++ src/app/(main)/layout.tsx
@@ -2,20 +2,13 @@
 import { ReactNode } from 'react';
 import '../styles/globals.css';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { useEffect } from 'react';
-import { useAuthStore } from '@/src/store/authStore';
-import Cookie from 'js-cookie';
+import { useAuthenticate } from '@/src/hooks/useAuthenticate';
 interface LayoutProps {
   children?: ReactNode;
 }
 
 export default function Layout({ children }: LayoutProps) {
-  const setIsLoggedIn = useAuthStore((state) => state.setIsLoggedIn);
-  useEffect(() => {
-    const loggedInStatus = Cookie.get('isLoggedIn') === 'true';
-    setIsLoggedIn(loggedInStatus);
-  }, [setIsLoggedIn]);
-
+  useAuthenticate();
   const queryClient = new QueryClient();
   return (
     <QueryClientProvider client={queryClient}>
